+++ demos/demo-events/src/demo_events/handlers/on_move_event.py
@@ -1,11 +1,11 @@
 from demo_events import models as models
-from demo_events.types.events_tzconst.event.unit import UnitPayload
+from demo_events.types.events_contract.event.move import MovePayload
 from dipdup.context import HandlerContext
 from dipdup.models import Event
 
 
-async def on_events_tzconst_unit(
+async def on_move_event(
     ctx: HandlerContext,
-    event: Event[UnitPayload],
+    event: Event[MovePayload],
 ) -> None:
     ...