+++ demos/demo-events/src/demo_events/handlers/on_other_event.py
@@ -3,7 +3,7 @@ from dipdup.context import HandlerContext
 from dipdup.models import UnknownEvent
 
 
-async def on_events_no_payload_unknown(
+async def on_other_event(
     ctx: HandlerContext,
     event: UnknownEvent,
 ) -> None: