+++ demos/demo-events/src/demo_events/handlers/on_roll_event.py
@@ -1,11 +1,11 @@
 from demo_events import models as models
-from demo_events.types.events_no_payload.event.option import OptionPayload
+from demo_events.types.events_contract.event.roll import RollPayload
 from dipdup.context import HandlerContext
 from dipdup.models import Event
 
 
-async def on_events_no_payload_option(
+async def on_roll_event(
     ctx: HandlerContext,
-    event: Event[OptionPayload],
+    event: Event[RollPayload],
 ) -> None:
     ...