+++ src/dipdup/context.py
@@ -337,6 +337,8 @@ class DipDupContext:
         # NOTE: IndexDispatcher will handle further initialization when it's time
         pending_indexes.append(index)
 
+    # TODO: disable_index(name: str)
+
     async def update_contract_metadata(
         self,
         network: str,
@@ -422,7 +424,7 @@ class DipDupContext:
 class HookContext(DipDupContext):
     """Execution context of hook callbacks.
 
-    :param hook_config: Configuration of current hook
+    :param hook_config: Configuration of the current hook
     """
 
     def __init__(
@@ -505,7 +507,7 @@ class TemplateValuesdict(dict[str, Any]):
 class HandlerContext(DipDupContext):
     """Execution context of handler callbacks.
 
-    :param handler_config: Configuration of current handler
+    :param handler_config: Configuration of the current handler
     :param datasource: Index datasource instance
     """
 
@@ -708,7 +710,7 @@ class CallbackManager:
 
     def _get_sql_path(self, ctx: 'DipDupContext', name: str) -> Path:
         subpackages = name.split('.')
-        sql_path = Path(ctx.config.package_path, 'sql', *subpackages)
+        sql_path = Path(env.get_package_path(ctx.config.package), 'sql', *subpackages)
         if not sql_path.exists():
             raise InitializationRequiredError(f'Missing SQL directory for hook `{name}`')