+++ src/dipdup/datasources/metadata/enums.py
@@ -4,7 +4,4 @@ from enum import Enum
 class MetadataNetwork(Enum):
     mainnet = 'mainnet'
     ghostnet = 'ghostnet'
-    # FIXME: Deactivated?
-    jakartanet = 'jakartanet'
-    kathmandunet = 'kathmandunet'
     limanet = 'limanet'